+++ src/gpio.rs
@@ -1,10 +1,68 @@
 //! # General Purpose I/Os
 //!
+//! The GPIO pins are organised into groups of 16 pins which can be accessed through the
+//! `gpioa`, `gpiob`... modules. To get access to the pins, you first need to convert them into a
+//! HAL designed struct from the `pac` struct using the `spilit` function.
+//! ```rust
+//! // Acquire the GPIOC peripheral
+//! // NOTE: `dp` is the device peripherals from the `PAC` crate
+//! let mut gpioa = dp.GPIOA.split(&mut rcc.apb2);
+//! ```
+//!
+//! This gives you a struct containing two control registers `crl` and `crh`, and all the pins
+//! `px0..px15`. These structs are what you use to interract with the pins to change their modes,
+//! or their inputs or outputs. For example, to set `pa5` high, you would call
+//!
+//! ```rust
+//! let output = gpioa.pa5.into_push_pull_output(&mut gpioa.crl);
+//! output.set_high();
+//! ```
+//!
+//! Each GPIO pin can be set to various modes:
+//!
+//! - **Alternate**: Pin mode required when the pin is driven by other peripherals
+//! - **Dynamic**: Pin mode is selected at runtime. See changing configurations for more details
+//! - Input
+//!     - **PullUp**: Input connected to high with a weak pull up resistor. Will be high when nothing
+//!     is connected
+//!     - **PullDown**: Input connected to high with a weak pull up resistor. Will be low when nothing
+//!     is connected
+//!     - **Floating**: Input not pulled to high or low. Will be undefined when nothing is connected
+//! - Output
+//!     - **PushPull**: Output which either drives the pin high or low
+//!     - **OpenDrain**: Output which leaves the gate floating, or pulls it do ground in drain
+//!     mode. Can be used as an input in the `open` configuration
+//! - **Debugger**: Some pins start out being used by the debugger. A pin in this mode can only be
+//! used if the [JTAG peripheral has been turned off](#accessing-pa15-pb3-and-pb14).
+//!
+//! ## Changing modes
+//! The simplest way to change the pin mode is to use the `into_<mode>` functions. These return a
+//! new struct with the correct mode that you can use the input or output functions on.
+//!
+//! If you need a more temporary mode change, and can not use the `into_<mode>` functions for
+//! ownership reasons, you can use the `as_<mode>` functions to temporarily change the pin type, do
+//! some output or input, and then have it change back once done.
+//!
+//! ### Dynamic Mode Change
+//! The above mode change methods guarantee that you can only call input functions when the pin is
+//! in input mode, and output when in output modes, but can lead to some issues. Therefore, there
+//! is also a mode where the state is kept track of at runtime, allowing you to change the mode
+//! often, and without problems with ownership, or references, at the cost of some performance and
+//! the risk of runtime errors.
+//!
+//! To make a pin dynamic, use the `into_dynamic` function, and then use the `make_<mode>` functions to
+//! change the mode
+//!
+//! ## Accessing PA15, PB3, and PB14
+//!
+//! These pins are used by the JTAG peripheral by default. To use them in your program, you need to
+//! disable that peripheral. This is done using the [afio::MAPR::disable_jtag](../afio/struct.MAPR.html#method.disable_jtag) function
+//!
 //! # Interfacing with v1 traits
 //!
-//! `embedded-hal` has two versions of the digital traits, `v2` which is used
-//! by this crate and `v1` which is deprecated but still used by a lot of drivers.
-//! If you want to use such a driver with this crate, you need to convert the digital pins to the `v1` type.
+//! `embedded-hal` has two versions of the digital traits, `v2` which is used by this crate and
+//! `v1` which is deprecated but still used by a lot of drivers.  If you want to use such a driver
+//! with this crate, you need to convert the digital pins to the `v1` type.
 //!
 //! This is done using `embedded-hal::digital::v1_compat::OldOutputPin`. For example:
 //!
@@ -12,7 +70,6 @@
 //! let nss = gpioa.pa4.into_push_pull_output(&mut gpioa.crl);
 //! let mut mfrc522 = Mfrc522::new(spi, OldOutputPin::from(nss)).unwrap();
 //! ```
-//!
 
 use core::marker::PhantomData;
 
@@ -96,6 +153,67 @@ pub trait ExtiPin {
     fn check_interrupt(&mut self) -> bool;
 }
 
+/// Tracks the current pin state for dynamic pins
+pub enum Dynamic {
+    InputFloating,
+    InputPullUp,
+    InputPullDown,
+    OutputPushPull,
+    OutputOpenDrain,
+}
+
+impl Active for Dynamic {}
+
+#[derive(Debug, PartialEq)]
+pub enum PinModeError {
+    IncorrectMode,
+}
+
+impl Dynamic {
+    fn is_input(&self) -> bool {
+        use Dynamic::*;
+        match self {
+            InputFloating | InputPullUp | InputPullDown | OutputOpenDrain => true,
+            OutputPushPull => false,
+        }
+    }
+    fn is_output(&self) -> bool {
+        use Dynamic::*;
+        match self {
+            InputFloating | InputPullUp | InputPullDown => false,
+            OutputPushPull | OutputOpenDrain => true,
+        }
+    }
+}
+
+/// NOTE: This trait should ideally be private but must be pub in order to avoid
+/// complaints from the compiler.
+pub trait PinMode<CR> {
+    unsafe fn set_mode(cr: &mut CR) -> Self;
+}
+
+// These impls are needed because a macro can not brace initialise a ty token
+impl<MODE> Input<MODE> {
+    const fn _new() -> Self {
+        Self { _mode: PhantomData }
+    }
+}
+impl<MODE> Output<MODE> {
+    const fn _new() -> Self {
+        Self { _mode: PhantomData }
+    }
+}
+impl<MODE> Alternate<MODE> {
+    const fn _new() -> Self {
+        Self { _mode: PhantomData }
+    }
+}
+impl Debugger {
+    const fn _new() -> Self {
+        Self {}
+    }
+}
+
 macro_rules! gpio {
     ($GPIOX:ident, $gpiox:ident, $gpioy:ident, $PXx:ident, $extigpionr:expr, [
         $($PXi:ident: ($pxi:ident, $i:expr, $MODE:ty, $CR:ident, $exticri:ident),)+
@@ -125,7 +243,10 @@ macro_rules! gpio {
                 Pxx,
                 Mode,
                 Edge,
-                ExtiPin
+                ExtiPin,
+                PinMode,
+                Dynamic,
+                PinModeError,
             };
 
             /// GPIO parts
@@ -151,7 +272,7 @@ macro_rules! gpio {
                         crl: CRL { _0: () },
                         crh: CRH { _0: () },
                         $(
-                            $pxi: $PXi { _mode: PhantomData },
+                            $pxi: $PXi { mode: <$MODE>::_new() },
                         )+
                     }
                 }
@@ -320,7 +441,7 @@ macro_rules! gpio {
             $(
                 /// Pin
                 pub struct $PXi<MODE> {
-                    _mode: PhantomData<MODE>,
+                    mode: MODE,
                 }
 
                 impl<MODE> Mode<MODE> for $PXi<MODE> {}
@@ -331,13 +452,14 @@ macro_rules! gpio {
                     /// state in the hardware.
                     #[allow(dead_code)]
                     pub(crate) unsafe fn activate(self) -> $PXi<Input<Floating>> {
-                        $PXi { _mode: PhantomData }
+                        $PXi { mode: Input::_new() }
                     }
                 }
 
                 impl<MODE> $PXi<MODE> where MODE: Active {
                     /// Configures the pin to operate as an alternate function push-pull output
                     /// pin.
+                    #[inline]
                     pub fn into_alternate_push_pull(
                         self,
                         cr: &mut $CR,
@@ -356,11 +478,12 @@ macro_rules! gpio {
                                 w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
                             });
 
-                        $PXi { _mode: PhantomData }
+                        $PXi { mode: Alternate::_new() }
                     }
 
                     /// Configures the pin to operate as an alternate function open-drain output
                     /// pin.
+                    #[inline]
                     pub fn into_alternate_open_drain(
                         self,
                         cr: &mut $CR,
@@ -379,85 +502,45 @@ macro_rules! gpio {
                                 w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
                             });
 
-                        $PXi { _mode: PhantomData }
+                        $PXi { mode: Alternate::_new() }
                     }
 
                     /// Configures the pin to operate as a floating input pin
+                    #[inline]
                     pub fn into_floating_input(
                         self,
                         cr: &mut $CR,
                     ) -> $PXi<Input<Floating>> {
-                        const OFFSET: u32 = (4 * $i) % 32;
-                        // Floating input
-                        const CNF: u32 = 0b01;
-                        // Input mode
-                        const MODE: u32 = 0b00;
-                        const BITS: u32 = (CNF << 2) | MODE;
-
-                        // input mode
-                        cr
-                            .cr()
-                            .modify(|r, w| unsafe {
-                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
-                            });
-
-                        $PXi { _mode: PhantomData }
+                        unsafe {
+                            $PXi::<Input<Floating>>::set_mode(cr)
+                        }
                     }
 
                     /// Configures the pin to operate as a pulled down input pin
+                    #[inline]
                     pub fn into_pull_down_input(
                         self,
                         cr: &mut $CR,
                     ) -> $PXi<Input<PullDown>> {
-                        const OFFSET: u32 = (4 * $i) % 32;
-                        // Pull up/down input
-                        const CNF: u32 = 0b10;
-                        // Input mode
-                        const MODE: u32 = 0b00;
-                        const BITS: u32 = (CNF << 2) | MODE;
-
-                        //pull down:
-                        // NOTE(unsafe) atomic write to a stateless register
-                        unsafe { (*$GPIOX::ptr()).bsrr.write(|w| w.bits(1 << (16 + $i))) }
-
-                        // input mode
-                        cr
-                            .cr()
-                            .modify(|r, w| unsafe {
-                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
-                            });
-
-                        $PXi { _mode: PhantomData }
+                        unsafe {
+                            $PXi::<Input<PullDown>>::set_mode(cr)
+                        }
                     }
 
                     /// Configures the pin to operate as a pulled up input pin
+                    #[inline]
                     pub fn into_pull_up_input(
                         self,
                         cr: &mut $CR,
                     ) -> $PXi<Input<PullUp>> {
-                        const OFFSET: u32 = (4 * $i) % 32;
-                        // Pull up/down input
-                        const CNF: u32 = 0b10;
-                        // Input mode
-                        const MODE: u32 = 0b00;
-                        const BITS: u32 = (CNF << 2) | MODE;
-
-                        //pull up:
-                        // NOTE(unsafe) atomic write to a stateless register
-                        unsafe { (*$GPIOX::ptr()).bsrr.write(|w| w.bits(1 << $i)) }
-
-                        // input mode
-                        cr
-                            .cr()
-                            .modify(|r, w| unsafe {
-                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
-                            });
-
-                        $PXi { _mode: PhantomData }
+                        unsafe {
+                            $PXi::<Input<PullUp>>::set_mode(cr)
+                        }
                     }
 
                     /// Configures the pin to operate as an open-drain output pin.
                     /// Initial state will be low.
+                    #[inline]
                     pub fn into_open_drain_output(
                         self,
                         cr: &mut $CR,
@@ -467,35 +550,20 @@ macro_rules! gpio {
 
                     /// Configures the pin to operate as an open-drain output pin.
                     /// `initial_state` specifies whether the pin should be initially high or low.
+                    #[inline]
                     pub fn into_open_drain_output_with_state(
-                        self,
+                        mut self,
                         cr: &mut $CR,
                         initial_state: State,
                     ) -> $PXi<Output<OpenDrain>> {
-                        const OFFSET: u32 = (4 * $i) % 32;
-                        // General purpose output open-drain
-                        const CNF: u32 = 0b01;
-                        // Open-Drain Output mode, max speed 50 MHz
-                        const MODE: u32 = 0b11;
-                        const BITS: u32 = (CNF << 2) | MODE;
-
-                        let mut res = $PXi { _mode: PhantomData };
-
-                        match initial_state {
-                            State::High => res.set_high(),
-                            State::Low  => res.set_low(),
-                        }.unwrap();
-
-                        cr
-                            .cr()
-                            .modify(|r, w| unsafe {
-                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
-                            });
-
-                        res
+                        self.set_state(initial_state);
+                        unsafe {
+                            $PXi::<Output<OpenDrain>>::set_mode(cr)
+                        }
                     }
                     /// Configures the pin to operate as an push-pull output pin.
                     /// Initial state will be low.
+                    #[inline]
                     pub fn into_push_pull_output(
                         self,
                         cr: &mut $CR,
@@ -505,60 +573,141 @@ macro_rules! gpio {
 
                     /// Configures the pin to operate as an push-pull output pin.
                     /// `initial_state` specifies whether the pin should be initially high or low.
+                    #[inline]
                     pub fn into_push_pull_output_with_state(
-                        self,
+                        mut self,
                         cr: &mut $CR,
                         initial_state: State,
                     ) -> $PXi<Output<PushPull>> {
-                        const OFFSET: u32 = (4 * $i) % 32;
-                        // General purpose output push-pull
-                        const CNF: u32 = 0b00;
-                        // Output mode, max speed 50 MHz
-                        const MODE: u32 = 0b11;
-                        const BITS: u32 = (CNF << 2) | MODE;
-
-                        let mut res = $PXi { _mode: PhantomData };
-
-                        match initial_state {
-                            State::High => res.set_high(),
-                            State::Low  => res.set_low(),
-                        }.unwrap();
-
-                        cr
-                            .cr()
-                            .modify(|r, w| unsafe {
-                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
-                            });
-
-                        res
+                        self.set_state(initial_state);
+                        unsafe {
+                            $PXi::<Output<PushPull>>::set_mode(cr)
+                        }
                     }
 
                     /// Configures the pin to operate as an analog input pin
+                    #[inline]
                     pub fn into_analog(self, cr: &mut $CR) -> $PXi<Analog> {
-                        const OFFSET: u32 = (4 * $i) % 32;
-                        // Analog input
-                        const CNF: u32 = 0b00;
-                        // Input mode
-                        const MODE: u32 = 0b00;
-                        const BITS: u32 = (CNF << 2) | MODE;
-
-                        // analog mode
-                        cr
-                            .cr()
-                            .modify(|r, w| unsafe {
-                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
-                            });
-
-                        $PXi { _mode: PhantomData }
+                        unsafe {
+                            $PXi::<Analog>::set_mode(cr)
+                        }
                     }
+
+                    /// Configures the pin as a pin that can change between input
+                    /// and output without changing the type. It starts out
+                    /// as a floating input
+                    #[inline]
+                    pub fn into_dynamic(self, cr: &mut $CR) -> $PXi<Dynamic> {
+                        self.into_floating_input(cr);
+                        $PXi::<Dynamic>{mode: Dynamic::InputFloating}
+                    }
+                }
+
+                // These macros are defined here instead of at the top level in order
+                // to be able to refer to macro variables from the outer layers.
+                macro_rules! impl_temp_output {
+                    (
+                        $fn_name:ident,
+                        $stateful_fn_name:ident,
+                        $mode:ty
+                    ) => {
+                        /**
+                          Temporarily change the mode of the pin.
+
+                          The value of the pin after conversion is undefined. If you
+                          want to control it, use `$stateful_fn_name`
+                        */
+                        #[inline]
+                        pub fn $fn_name(
+                            &mut self,
+                            cr: &mut $CR,
+                            mut f: impl FnMut(&mut $PXi<$mode>)
+                        ) {
+                            let mut temp = unsafe { $PXi::<$mode>::set_mode(cr) };
+                            f(&mut temp);
+                            unsafe {
+                                Self::set_mode(cr);
+                            }
+                        }
+
+                        /**
+                          Temporarily change the mode of the pin.
+
+                          Note that the new state is set slightly before conversion
+                          happens. This can cause a short output glitch if switching
+                          between output modes
+                        */
+                        #[inline]
+                        pub fn $stateful_fn_name(
+                            &mut self,
+                            cr: &mut $CR,
+                            state: State,
+                            mut f: impl FnMut(&mut $PXi<$mode>)
+                        ) {
+                            self.set_state(state);
+                            let mut temp = unsafe { $PXi::<$mode>::set_mode(cr) };
+                            f(&mut temp);
+                            unsafe {
+                                Self::set_mode(cr);
+                            }
+                        }
+                    }
+                }
+                macro_rules! impl_temp_input {
+                    (
+                        $fn_name:ident,
+                        $mode:ty
+                    ) => {
+                        /**
+                          Temporarily change the mode of the pin.
+                        */
+                        #[inline]
+                        pub fn $fn_name(
+                            &mut self,
+                            cr: &mut $CR,
+                            mut f: impl FnMut(&mut $PXi<$mode>)
+                        ) {
+                            let mut temp = unsafe { $PXi::<$mode>::set_mode(cr) };
+                            f(&mut temp);
+                            unsafe {
+                                Self::set_mode(cr);
+                            }
+                        }
+                    }
+                }
+
+                impl<MODE> $PXi<MODE> where MODE: Active, $PXi<MODE>: PinMode<$CR> {
+                    impl_temp_output!(
+                        as_push_pull_output,
+                        as_push_pull_output_with_state,
+                        Output<PushPull>
+                    );
+                    impl_temp_output!(
+                        as_open_drain_output,
+                        as_open_drain_output_with_state,
+                        Output<OpenDrain>
+                    );
+                    impl_temp_input!(
+                        as_floating_input,
+                        Input<Floating>
+                    );
+                    impl_temp_input!(
+                        as_pull_up_input,
+                        Input<PullUp>
+                    );
+                    impl_temp_input!(
+                        as_pull_down_input,
+                        Input<PullDown>
+                    );
                 }
 
                 impl<MODE> $PXi<MODE> where MODE: Active {
                     /// Erases the pin number from the type
+                    #[inline]
                     fn into_generic(self) -> Generic<MODE> {
                         Generic {
                             i: $i,
-                            _mode: self._mode,
+                            _mode: PhantomData,
                         }
                     }
 
@@ -571,27 +720,32 @@ macro_rules! gpio {
                     }
                 }
 
+                // embedded_hal impls
+
                 impl<MODE> OutputPin for $PXi<Output<MODE>> {
                     type Error = Infallible;
+                    #[inline]
                     fn set_high(&mut self) -> Result<(), Self::Error> {
                         // NOTE(unsafe) atomic write to a stateless register
-                        Ok(unsafe { (*$GPIOX::ptr()).bsrr.write(|w| w.bits(1 << $i)) })
+                        Ok(self.set_state(State::High))
                     }
 
+                    #[inline]
                     fn set_low(&mut self) -> Result<(), Self::Error> {
                         // NOTE(unsafe) atomic write to a stateless register
-                        Ok(unsafe { (*$GPIOX::ptr()).bsrr.write(|w| w.bits(1 << (16 + $i))) })
+                        Ok(self.set_state(State::Low))
                     }
                 }
 
                 impl<MODE> StatefulOutputPin for $PXi<Output<MODE>> {
+                    #[inline]
                     fn is_set_high(&self) -> Result<bool, Self::Error> {
                         self.is_set_low().map(|b| !b)
                     }
 
+                    #[inline]
                     fn is_set_low(&self) -> Result<bool, Self::Error> {
-                        // NOTE(unsafe) atomic read with no side effects
-                        Ok(unsafe { (*$GPIOX::ptr()).odr.read().bits() & (1 << $i) == 0 })
+                        Ok(self._is_set_low())
                     }
                 }
 
@@ -599,27 +753,106 @@ macro_rules! gpio {
 
                 impl<MODE> InputPin for $PXi<Input<MODE>> {
                     type Error = Infallible;
+                    #[inline]
                     fn is_high(&self) -> Result<bool, Self::Error> {
                         self.is_low().map(|b| !b)
                     }
 
+                    #[inline]
                     fn is_low(&self) -> Result<bool, Self::Error> {
                         // NOTE(unsafe) atomic read with no side effects
-                        Ok(unsafe { (*$GPIOX::ptr()).idr.read().bits() & (1 << $i) == 0 })
+                        Ok(self._is_low())
                     }
                 }
 
                 impl InputPin for $PXi<Output<OpenDrain>> {
                     type Error = Infallible;
+                    #[inline]
                     fn is_high(&self) -> Result<bool, Self::Error> {
                         self.is_low().map(|b| !b)
                     }
 
+                    #[inline]
                     fn is_low(&self) -> Result<bool, Self::Error> {
-                        // NOTE(unsafe) atomic read with no side effects
-                        Ok(unsafe { (*$GPIOX::ptr()).idr.read().bits() & (1 << $i) == 0 })
+                        Ok(self._is_low())
                     }
                 }
+
+
+                // Dynamic pin
+
+                impl $PXi<Dynamic> {
+                    #[inline]
+                    pub fn make_pull_up_input(&mut self, cr: &mut $CR) {
+                        // NOTE(unsafe), we have a mutable reference to the current pin
+                        unsafe { $PXi::<Input<PullUp>>::set_mode(cr) };
+                        self.mode = Dynamic::InputPullUp;
+                    }
+                    #[inline]
+                    pub fn make_pull_down_input(&mut self, cr: &mut $CR) {
+                        // NOTE(unsafe), we have a mutable reference to the current pin
+                        unsafe { $PXi::<Input<PullDown>>::set_mode(cr) };
+                        self.mode = Dynamic::InputPullDown;
+                    }
+                    #[inline]
+                    pub fn make_floating_input(&mut self, cr: &mut $CR) {
+                        // NOTE(unsafe), we have a mutable reference to the current pin
+                        unsafe { $PXi::<Input<Floating>>::set_mode(cr) };
+                        self.mode = Dynamic::InputFloating;
+                    }
+                    #[inline]
+                    pub fn make_push_pull_output(&mut self, cr: &mut $CR) {
+                        // NOTE(unsafe), we have a mutable reference to the current pin
+                        unsafe { $PXi::<Output<PushPull>>::set_mode(cr) };
+                        self.mode = Dynamic::OutputPushPull;
+                    }
+                    #[inline]
+                    pub fn make_open_drain_output(&mut self, cr: &mut $CR) {
+                        // NOTE(unsafe), we have a mutable reference to the current pin
+                        unsafe { $PXi::<Output<OpenDrain>>::set_mode(cr) };
+                        self.mode = Dynamic::OutputOpenDrain;
+                    }
+                }
+
+                impl OutputPin for $PXi<Dynamic> {
+                    type Error = PinModeError;
+                    fn set_high(&mut self) -> Result<(), Self::Error> {
+                        if self.mode.is_output() {
+                            self.set_state(State::High);
+                            Ok(())
+                        }
+                        else {
+                            Err(PinModeError::IncorrectMode)
+                        }
+                    }
+                    fn set_low(&mut self) -> Result<(), Self::Error> {
+                        if self.mode.is_output() {
+                            self.set_state(State::Low);
+                            Ok(())
+                        }
+                        else {
+                            Err(PinModeError::IncorrectMode)
+                        }
+                    }
+                }
+
+                impl InputPin for $PXi<Dynamic> {
+                    type Error = PinModeError;
+                    fn is_high(&self) -> Result<bool, Self::Error> {
+                        self.is_low().map(|b| !b)
+                    }
+                    fn is_low(&self) -> Result<bool, Self::Error> {
+                        if self.mode.is_input() {
+                            Ok(self._is_low())
+                        }
+                        else {
+                            Err(PinModeError::IncorrectMode)
+                        }
+                    }
+                }
+
+
+                // Exti pin impls
 
                 impl<MODE> ExtiPin for $PXi<Input<MODE>> {
                     /// Configure EXTI Line $i to trigger from this pin.
@@ -651,23 +884,189 @@ macro_rules! gpio {
                     }
 
                     /// Enable external interrupts from this pin.
+                    #[inline]
                     fn enable_interrupt(&mut self, exti: &EXTI) {
                         exti.imr.modify(|r, w| unsafe { w.bits(r.bits() | (1 << $i)) });
                     }
 
                     /// Disable external interrupts from this pin
+                    #[inline]
                     fn disable_interrupt(&mut self, exti: &EXTI) {
                         exti.imr.modify(|r, w| unsafe { w.bits(r.bits() & !(1 << $i)) });
                     }
 
                     /// Clear the interrupt pending bit for this pin
+                    #[inline]
                     fn clear_interrupt_pending_bit(&mut self) {
                         unsafe { (*EXTI::ptr()).pr.write(|w| w.bits(1 << $i) ) };
                     }
 
                     /// Reads the interrupt pending bit for this pin
+                    #[inline]
                     fn check_interrupt(&mut self) -> bool {
                         unsafe { ((*EXTI::ptr()).pr.read().bits() & (1 << $i)) != 0 }
+                    }
+                }
+
+
+                // Internal helper functions
+
+                // NOTE: The functions in this impl block are "safe", but they
+                // are callable when the pin is in modes where they don't make
+                // sense.
+                impl<MODE> $PXi<MODE> {
+                    /**
+                      Set the output of the pin regardless of its mode.
+                      Primarily used to set the output value of the pin
+                      before changing its mode to an output to avoid
+                      a short spike of an incorrect value
+                    */
+                    fn set_state(&mut self, state: State) {
+                        match state {
+                            State::High => unsafe {
+                                (*$GPIOX::ptr()).bsrr.write(|w| w.bits(1 << $i))
+                            }
+                            State::Low => unsafe {
+                                (*$GPIOX::ptr()).bsrr.write(|w| w.bits(1 << (16 + $i)))
+                            }
+                        }
+                    }
+
+                    fn _is_set_low(&self) -> bool {
+                        unsafe { (*$GPIOX::ptr()).odr.read().bits() & (1 << $i) == 0 }
+                    }
+
+                    fn _is_low(&self) -> bool {
+                        // NOTE(unsafe) atomic read with no side effects
+                        unsafe { (*$GPIOX::ptr()).idr.read().bits() & (1 << $i) == 0 }
+                    }
+                }
+
+                impl PinMode<$CR> for $PXi<Input<Floating>> {
+                    unsafe fn set_mode(cr: &mut $CR) -> Self {
+                        const OFFSET: u32 = (4 * $i) % 32;
+                        // Floating input
+                        const CNF: u32 = 0b01;
+                        // Input mode
+                        const MODE: u32 = 0b00;
+                        const BITS: u32 = (CNF << 2) | MODE;
+
+                        // input mode
+                        cr
+                            .cr()
+                            .modify(|r, w| {
+                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
+                            });
+
+                        $PXi { mode: Input::_new() }
+                    }
+                }
+
+                impl PinMode<$CR> for $PXi<Input<PullDown>> {
+                    unsafe fn set_mode(cr: &mut $CR) -> Self {
+                        const OFFSET: u32 = (4 * $i) % 32;
+                        // Pull up/down input
+                        const CNF: u32 = 0b10;
+                        // Input mode
+                        const MODE: u32 = 0b00;
+                        const BITS: u32 = (CNF << 2) | MODE;
+
+                        //pull down:
+                        // NOTE(unsafe) atomic write to a stateless register
+                        (*$GPIOX::ptr()).bsrr.write(|w| w.bits(1 << (16 + $i)));
+
+                        // input mode
+                        cr
+                            .cr()
+                            .modify(|r, w| {
+                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
+                            });
+
+                        $PXi { mode: Input::_new() }
+                    }
+                }
+
+
+                impl PinMode<$CR> for $PXi<Input<PullUp>> {
+                    unsafe fn set_mode(cr: &mut $CR) -> Self {
+                        const OFFSET: u32 = (4 * $i) % 32;
+                        // Pull up/down input
+                        const CNF: u32 = 0b10;
+                        // Input mode
+                        const MODE: u32 = 0b00;
+                        const BITS: u32 = (CNF << 2) | MODE;
+
+                        //pull up:
+                        // NOTE(unsafe) atomic write to a stateless register
+                        (*$GPIOX::ptr()).bsrr.write(|w| w.bits(1 << $i));
+
+                        // input mode
+                        cr
+                            .cr()
+                            .modify(|r, w| {
+                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
+                            });
+
+                        $PXi { mode: Input::_new() }
+                    }
+                }
+
+                impl PinMode<$CR> for $PXi<Output<OpenDrain>> {
+                    unsafe fn set_mode(cr: &mut $CR) -> Self {
+                        const OFFSET: u32 = (4 * $i) % 32;
+                        // General purpose output open-drain
+                        const CNF: u32 = 0b01;
+                        // Open-Drain Output mode, max speed 50 MHz
+                        const MODE: u32 = 0b11;
+                        const BITS: u32 = (CNF << 2) | MODE;
+
+                        cr
+                            .cr()
+                            .modify(|r, w| {
+                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
+                            });
+
+                        $PXi { mode: Output::_new() }
+                    }
+                }
+
+                impl PinMode<$CR> for $PXi<Output<PushPull>> {
+                    unsafe fn set_mode(cr: &mut $CR) -> Self {
+                        const OFFSET: u32 = (4 * $i) % 32;
+                        // General purpose output push-pull
+                        const CNF: u32 = 0b00;
+                        // Output mode, max speed 50 MHz
+                        const MODE: u32 = 0b11;
+                        const BITS: u32 = (CNF << 2) | MODE;
+
+
+                        cr
+                            .cr()
+                            .modify(|r, w| {
+                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
+                            });
+
+                        $PXi { mode: Output::_new() }
+                    }
+                }
+
+                impl PinMode<$CR> for $PXi<Analog> {
+                    unsafe fn set_mode(cr: &mut $CR) -> Self {
+                        const OFFSET: u32 = (4 * $i) % 32;
+                        // Analog input
+                        const CNF: u32 = 0b00;
+                        // Input mode
+                        const MODE: u32 = 0b00;
+                        const BITS: u32 = (CNF << 2) | MODE;
+
+                        // analog mode
+                        cr
+                            .cr()
+                            .modify(|r, w| {
+                                w.bits((r.bits() & !(0b1111 << OFFSET)) | (BITS << OFFSET))
+                            });
+
+                        $PXi { mode: Analog{} }
                     }
                 }
             )+