+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Add runtime-reconfigurable GPIO pins
+
 ### Fixed
 
 - Fix period retrieval for timers